requirements.txt
django==1.3
django-gravatar
Markdown
